nouvelles fonctions

+++ projetJava/src/DaoBase.java
@@ -1,17 +1,50 @@
-package WebServer;
+package projet;
 
 import java.util.Date;
 import java.util.List;
 import java.util.ListIterator;
 
+
 import org.sql2o.Connection;
 import org.sql2o.Sql2o;
 
 public class DaoBase {
-	private static String connection = "jdbc:h2:~/test;";
-	private static Sql2o sql2o = new Sql2o(connection, "Clement", "");
+	private static  String connection = "jdbc:h2:~/test;";
+	private static  Sql2o sql2o = new Sql2o(connection, "Clement", "");
 	
-	public void insertListe(Liste liste){
+	/*
+	public DaoBase() {
+		final String insertQuery="DROP TABLE IF EXISTS LISTE;" + 
+				"DROP TABLE IF EXISTS ELEMENT;" + 
+				"DROP TABLE IF EXISTS ASSOCIATION;" + 
+				"CREATE TABLE LISTE(ID UUID PRIMARY KEY, TITRE VARCHAR(255), DESCRIPTION VARCHAR(255));" + 
+				"CREATE TABLE ELEMENT(ID UUID PRIMARY KEY, DATECREA DATE, DATEMODIF DATE, TITRE VARCHAR(255), DESCRIPTION VARCHAR(255));" + 
+				"CREATE TABLE ASSOCIATION(IDLISTE UUID, IDELEM UUID,TYPE BIT,PRIMARY KEY (IDLISTE,IDELEM));";
+		try(Connection connect = sql2o.beginTransaction()){
+			 connect.createQuery(insertQuery)
+			 	.executeUpdate();
+			 connect.commit();
+		}
+	}
+	
+	*/
+ 	public static  boolean verifCaract(String str ) {
+		return str.matches("[a-zA-Z_0-9-]*");
+	}
+	
+ 	public void EffacerTout(){
+		
+		final String insertQuery="DELETE  FROM ASSOCIATION; DELETE FROM ELEMENT; DELETE FROM LISTE";
+		try(Connection connect = sql2o.beginTransaction()){
+			 connect.createQuery(insertQuery)
+		        .executeUpdate();
+			 connect.commit();
+		}
+	}
+ 	
+	public void insert(Liste liste){
+		if (verifCaract(liste.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de la liste");
+		if (verifCaract(liste.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans description liste");
 		final String insertQuery="INSERT INTO LISTE(ID,TITRE,DESCRIPTION) "+ "VALUES(:ID, :TITRE, :DESCRIPTION)";
 		try(Connection connect = sql2o.beginTransaction()){
 			 connect.createQuery(insertQuery)
@@ -23,9 +56,40 @@ public class DaoBase {
 		}
 	}
 	
-	public void insertElement(Element element, Liste liste){ // modif param de idliste Ã  liste
-		final String insertQuery="INSERT INTO ELEMENT(ID,TITRE,DESCRIPTION,DATECREA,DATEMODIF) "+ "VALUES(:ID, :TITRE, :DESCRIPTION, :DATECREA, :DATEMODIF)";
-		final String insertQuery2="INSERT INTO ASSOCIATION(IDLISTE,IDELEM) "+ "VALUES(:IDLISTE,:IDELEM)";
+	public void insert(Liste listeInsert, Liste listeRecept){ // modif param de idliste à liste
+		if (verifCaract(listeInsert.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre");
+		if (verifCaract(listeInsert.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le tritre");
+		if (verifCaract(listeRecept.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de la liste");
+		if (verifCaract(listeRecept.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans description liste");
+		if (listeRecept.getId()==listeInsert.getId()) throw new IllegalArgumentException("Même Id pour les deux listes");
+		
+		final String insertQuery="INSERT INTO LISTE(ID,TITRE,DESCRIPTION) "+ "VALUES(:ID, :TITRE, :DESCRIPTION) "+ "VALUES(:ID, :TITRE, :DESCRIPTION)";
+		final String insertQuery2="INSERT INTO ASSOCIATION(IDLISTE,IDELEM) "+ "VALUES(:IDLISTE,:IDELEM)" + "VALUES(:TYPE,:type)";
+		try(Connection connect = sql2o.beginTransaction()){
+			 connect.createQuery(insertQuery)
+		        .addParameter("ID", listeInsert.getId())
+		        .addParameter("TITRE", listeInsert.getTitre())
+		        .addParameter("DESCRIPTION", listeInsert.getDescrip())		        
+		        		        
+		        .executeUpdate();
+			 connect.createQuery(insertQuery2)  //ajout du lien dans la table Association
+		        .addParameter("IDLISTE", listeRecept.getId())
+		        .addParameter("IDELEM", listeRecept.getId())
+		        .addParameter("type", 0)
+		        .executeUpdate();
+			 connect.commit();
+		}
+	}
+	
+	
+	public void insert(Element element, Liste liste){ // modif param de idliste à liste
+		if (verifCaract(element.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre");
+		if (verifCaract(element.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le tritre");
+		if (verifCaract(liste.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de la liste");
+		if (verifCaract(liste.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans description liste");
+		
+		final String insertQuery="INSERT INTO ELEMENT(ID,DATECREA,DATEMODIF,TITRE,DESCRIPTION) "+ "VALUES(:ID, :DATECREA, :DATEMODIF, :TITRE, :DESCRIPTION)";
+		final String insertQuery2="INSERT INTO ASSOCIATION(IDLISTE,IDELEM,TYPE) "+ "VALUES(:IDLISTE,:IDELEM,:type)";
 		try(Connection connect = sql2o.beginTransaction()){
 			 connect.createQuery(insertQuery)
 		        .addParameter("ID", element.getId())
@@ -36,7 +100,8 @@ public class DaoBase {
 		        .executeUpdate();
 			 connect.createQuery(insertQuery2)  //ajout du lien dans la table Association
 		        .addParameter("IDLISTE", liste.getId())
-		        .addParameter("IDELEM", element.getId())		        
+		        .addParameter("IDELEM", element.getId())
+		        .addParameter("type", 1)
 		        .executeUpdate();
 			 connect.commit();
 		
@@ -44,45 +109,85 @@ public class DaoBase {
 		
 	}
 	
-	public void modifElement(Element element){
-		final String insertQuery="UPDATE ELEMENT SET TITRE=:TITRE,DESCRIPTION=:DESCRIPTION,DATEMODIF=:DATEMODIF WHERE IDELEM=:ID";
+	public void modif(String Idelem, Element element){
+		if (verifCaract(element.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de l'elem");
+		if (verifCaract(element.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le tritre de l'elem");
+		final String insertQuery="UPDATE ELEMENT SET TITRE=:TITRE,DESCRIPTION=:DESCRIPTION,DATEMODIF=:DATEMODIF WHERE ID=:Idelem";
 		try(Connection connect = sql2o.beginTransaction()){
 			 connect.createQuery(insertQuery)
 		        .addParameter("TITRE", element.getTitre())
 		        .addParameter("DESCRIPTION", element.getDescrip())		        
-		        .addParameter("DATEMODIF", new Date())		//problÃ¨me on fait une modification sans passer par setTitre et setDescripton donc DateModif est pas changÃ©e        
-		        .addParameter("ID", element.getId())
+		        .addParameter("DATEMODIF", new Date())		//problème on fait une modification sans passer par setTitre et setDescripton donc DateModif est pas changée        
+		        .addParameter("Idelem", Idelem)
 		        .executeUpdate();
 			 connect.commit();
 		}
 	}
 	
-	public void modifListe(Liste liste){
-		final String insertQuery="UPDATE LISTE SET TITRE=:TITRE,DESCRIPTION=:DESCRIPTION WHERE IDLISTE=:ID";
+	public void modif(String IdListeModif,Liste liste){
+		if (verifCaract(IdListeModif)== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre");
+		if (verifCaract(liste.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre");
+		if (verifCaract(liste.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le tritre");
+		final String insertQuery="UPDATE LISTE SET TITRE=:TITRE,DESCRIPTION=:DESCRIPTION WHERE ID=:ID";
 		try(Connection connect = sql2o.beginTransaction()){
 			 connect.createQuery(insertQuery)
 		        .addParameter("TITRE", liste.getTitre())
 		        .addParameter("DESCRIPTION", liste.getDescrip())		        
-		        .addParameter("ID", liste.getId())
+		        .addParameter("ID", IdListeModif)
 		        .executeUpdate();
 			 connect.commit();
 		}
 	}
 	
-	public void supprElementDansListe(Element element,Liste liste) {
-		final String insertQuery="DELETE  FROM ASSOCIATION WHERE IDELEM=:idelem AND IDLIST=: idlist;";
+	
+	public void supprimer(Element element,Liste liste) {
+		if (verifCaract(element.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de l'elem");
+		if (verifCaract(element.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le tritre de l'elem");
+		if (verifCaract(liste.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de la liste");
+		if (verifCaract(liste.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans description liste");
+		
+		final String insertQuery="DELETE  FROM ASSOCIATION WHERE IDELEM=:IDE AND IDLISTE=:IDL;";
 		try(Connection connect = sql2o.beginTransaction()){
 			 connect.createQuery(insertQuery)	        
-		        .addParameter("idelem", element.getId())
-		        .addParameter("idlist", liste.getId())
+		        .addParameter("IDE", element.getId())
+		        .addParameter("IDL", liste.getId())
 		        .executeUpdate();
 			 connect.commit();
 		}
-		if (elementSansListe(element)) supprElement(element)  ;
+		if (elementSansListe(element)) {
+			final String insertQuery2="DELETE FROM ELEMENT WHERE ID=:ID;";
+			try(Connection connect = sql2o.beginTransaction()){
+				 connect.createQuery(insertQuery2)	        
+			        .addParameter("ID", element.getId())
+			        .executeUpdate();
+				 connect.commit();
+			}
+		}
+	}
+	public void supprimer(Liste sousliste,Liste liste) {
+		if (verifCaract(sousliste.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de l'elem");
+		if (verifCaract(sousliste.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le tritre de l'elem");
+		if (verifCaract(liste.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de la liste");
+		if (verifCaract(liste.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans description liste");
+		
+		final String insertQuery="DELETE  FROM ASSOCIATION WHERE IDELEM=: IDE AND IDLISTE=: IDL AND TYPE=0;";
+		try(Connection connect = sql2o.beginTransaction()){
+			 connect.createQuery(insertQuery)	        
+		        .addParameter("IDE", sousliste.getId())
+		        .addParameter("IDL", liste.getId())
+		        .executeUpdate();
+			 connect.commit();
+		}
+			
+		supprimer(sousliste);
+			
 		
 	}
 	
-	public void supprElement(Element element){
+	public void supprimer(Element element){
+		if (verifCaract(element.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de l'elem");
+		if (verifCaract(element.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le tritre de l'elem");
+		
 		final String insertQuery="DELETE FROM ELEMENT WHERE ID=:ID;";
 		try(Connection connect = sql2o.beginTransaction()){
 			 connect.createQuery(insertQuery)	        
@@ -100,14 +205,22 @@ public class DaoBase {
 		}
 	}
 	
-	public void supprListe(Liste liste){
+	public void supprimer(Liste liste){
+		if (verifCaract(liste.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de la liste");
+		if (verifCaract(liste.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans description liste");
 		ListIterator<Element> li = getElementsPerList(liste).listIterator();  
 		Element element;
-		while(li.hasNext()) {      
-			element = li.next();
-			supprElementDansListe(element,liste);
+		while(li.hasNext()) {
+				 element = li.next();
+				 supprimer(element,liste);
 		}
 		
+		ListIterator<Liste> liL = getListPerList(liste).listIterator();  
+		Liste sousListe;	
+		while(liL.hasNext()) {
+			 sousListe = liL.next();
+			 supprimer(sousListe);
+		}		 
 		final String insertQuery="DELETE FROM LISTE WHERE ID=:ID;";
 		try(Connection connect = sql2o.beginTransaction()){
 			 connect.createQuery(insertQuery)	        
@@ -117,6 +230,7 @@ public class DaoBase {
 		}
 		
 	}
+	
 	
 	public static List<Liste> getAllLists(){
 		
@@ -128,19 +242,9 @@ public class DaoBase {
 		        .executeAndFetch(Liste.class);
 		  }
 	}
-	
-	public static List<Liste> getListsByID(String id){
-		
-		try (Connection con = sql2o.open()) {
-		    final String query =
-		        "SELECT * FROM LISTE WHERE ID= :idlist";
-		    return con.createQuery(query)
-		    		.addParameter("idlist",id)
-		    		.executeAndFetch(Liste.class);
-		  }
-	}
-	
-	public static List<Liste> getListsByTitre(String titre){
+
+	public  List<Liste> getListsByTitre(String titre){
+		if (verifCaract(titre)== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre ");;
 		
 		try (Connection con = sql2o.open()) {
 		    final String query =
@@ -152,7 +256,9 @@ public class DaoBase {
 		  }
 	}
 	
-	public static List<Liste> getListsByDescription(String description){
+	public  List<Liste> getListsByDescription(String description){
+		
+		if (verifCaract(description)== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans description liste");
 		
 		try (Connection con = sql2o.open()) {
 		    final String query =
@@ -164,23 +270,39 @@ public class DaoBase {
 		  }
 	}
 	
-	public static List<Element> getElementsPerList(Liste list){ 
-		
+	
+	public  List<Element> getElementsPerList(Liste list){ 
+		if (verifCaract(list.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre");
+		if (verifCaract(list.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le tritre");
 		try (Connection con = sql2o.open()) {
 		    final String query =
-		        "SELECT * FROM ELEMENT WHERE ID IN(SELECT IDELEM FROM ASSOCIATION WHERE IDLISTE =:idliste)";
+		        "SELECT * FROM ELEMENT WHERE ID IN(SELECT IDELEM FROM ASSOCIATION WHERE IDLISTE =:idliste AND TYPE=1)";
 		    		
 		    return con.createQuery(query)
 	    		.addParameter("idliste", list.getId())
 		        .executeAndFetch(Element.class);
 		  }
 	}
+	public  List<Liste> getListPerList(Liste list){ 
+		if (verifCaract(list.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre");
+		if (verifCaract(list.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le tritre");
+		try (Connection con = sql2o.open()) {
+		    final String query =
+		        "SELECT * FROM LISTE WHERE ID IN(SELECT IDELEM FROM ASSOCIATION WHERE IDLISTE =:idliste AND TYPE=0)";
+		    		
+		    return con.createQuery(query)
+	    		.addParameter("idliste", list.getId())
+		        .executeAndFetch(Liste.class);
+		  }
+	}
 	
-	public static boolean elementSansListe(Element element){
+	public  boolean elementSansListe(Element element){
 		return (getListPerElement(element).size()==0);
 	}
 	
-	public static List<Liste> getListPerElement(Element element){ 
+	public  List<Liste> getListPerElement(Element element){ 
+		if (verifCaract(element.getTitre() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre de l'elem");
+		if (verifCaract(element.getDescrip() )== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans descript de l'elem");
 		
 		try (Connection con = sql2o.open()) {
 		    final String query =
@@ -192,40 +314,45 @@ public class DaoBase {
 		  }
 	}
 	
-	public static List<Element> getAllElements(){  
+	public  List<Element> getAllElements(){  
 		
 		try (Connection con = sql2o.open()) {
 		    final String query =
-		        "SELECT  * FROM ELEMENT"; // *  Ã  la place des champs ?
+		        "SELECT  * FROM ELEMENT"; // *  à la place des champs ?
 
 		    return con.createQuery(query)
 		        .executeAndFetch(Element.class);
 		  }
 	}
 	
-	public static List<Element> getElementsByCreateDate(Date dateCrea){
+	public  List<Element> getElementsByCreateDate(Date dateCrea){
 		
 		try (Connection con = sql2o.open()) {
 		    final String query =
 		        "SELECT  * FROM ELEMENT WHERE DATECREA = :dateCrea";
 
 		    return con.createQuery(query)
-		        .addParameter("DATECREA", dateCrea).executeAndFetch(Element.class);
+		        .addParameter("DATECREA", dateCrea)
+		        .executeAndFetch(Element.class);
 		  }
 	}
 	
-	public static List<Element> getElementsByModifDate(Date dateModif){
+	public  List<Element> getElementsByModifDate(Date dateModif){
 		
 		try (Connection con = sql2o.open()) {
 		    final String query =
 		        "SELECT  * FROM ELEMENT WHERE DATEMODIF = :dateModif";
 
 		    return con.createQuery(query)
-		        .addParameter("DATEMODIF", dateModif).executeAndFetch(Element.class);
+		        .addParameter("DATEMODIF", dateModif)
+		        .executeAndFetch(Element.class);
+		    	
 		  }
 	}
 	
 	public static List<Element> getElementsByTitre(String titre){
+		if (verifCaract(titre)== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans le titre");
+		
 		
 		try (Connection con = sql2o.open()) {
 		    final String query =
@@ -237,26 +364,27 @@ public class DaoBase {
 		  }
 	}
 	
-	public static List<Element> getElementsByID(String id){
+	public static List<Element> getElementsByID(String Idelem){
+		if (verifCaract(Idelem)== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans param");
 		
 		try (Connection con = sql2o.open()) {
 		    final String query =
 		        "SELECT * FROM ELEMENT WHERE ID= :idelem";
 		    return con.createQuery(query)
-		    		.addParameter("idelem",id)
+		    		.addParameter("idelem",Idelem)
 		    		.executeAndFetch(Element.class);
+		  }
+	}
+
+	public static List<Liste> getListesByID(String Idliste){
+		if (verifCaract(Idliste)== false)  throw new IllegalArgumentException("caractère spéciale utilisé dans param");
+		try (Connection con = sql2o.open()) {
+		    final String query =
+		        "SELECT * FROM LISTE WHERE ID= :idliste";
+		    return con.createQuery(query)
+		    		.addParameter("idliste",Idliste)
+		    		.executeAndFetch(Liste.class);
 		  }
 	}
 	
-	public static List<Element> getElementsByDescription(String description){
-		
-		try (Connection con = sql2o.open()) {
-		    final String query =
-		        "SELECT * FROM ELEMENT WHERE DESCRIPTION = :description";
-
-		    return con.createQuery(query)
-		    		.addParameter("description", description)
-		    		.executeAndFetch(Element.class);
-		  }
-	}
 }
